LayoutTests/storage/indexeddb/readonly-properties-expected.txt

Test IndexedDB readonly properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
PASS Boolean(indexedDB && IDBCursor && IDBDatabase && IDBDatabaseError && IDBDatabaseException && IDBFactory && IDBIndex && IDBKeyRange && IDBObjectStore && IDBRequest && IDBTransaction) is true

indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
transaction = event.target.result;
Deleted all object stores.
objectStore = db.createObjectStore('foo');
trying to set readonly property objectStore.transaction
objectStore.transaction = this
PASS objectStore.transaction is still [object IDBTransaction]
PASS successfullyParsed is true

TEST COMPLETE
